Removed api project because it was giving migration errors. added identical project without errors
