+++ stuffWebApi/stuffWebApi.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2020
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "stuffioAPI", "stuffioAPI\stuffioAPI.csproj", "{073BBF27-63D3-446A-BDD2-1245E3920D39}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "stuffWebApi", "stuffWebApi\stuffWebApi.csproj", "{40969CAD-26BC-4705-BA4B-4E82C5D8C87E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{073BBF27-63D3-446A-BDD2-1245E3920D39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{073BBF27-63D3-446A-BDD2-1245E3920D39}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{073BBF27-63D3-446A-BDD2-1245E3920D39}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{073BBF27-63D3-446A-BDD2-1245E3920D39}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40969CAD-26BC-4705-BA4B-4E82C5D8C87E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40969CAD-26BC-4705-BA4B-4E82C5D8C87E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40969CAD-26BC-4705-BA4B-4E82C5D8C87E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40969CAD-26BC-4705-BA4B-4E82C5D8C87E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {98FE5A54-4D84-4082-9C60-3128CBE03648}
+		SolutionGuid = {B91139E1-F19A-4323-9D20-22711CB9E557}
 	EndGlobalSection
 EndGlobal
