+++ stuffWebApi/stuffWebApi/Models/Chat.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace stuffioAPI.Models
+namespace stuffWebApi.Models
 {
     public class Chat
     {
