+++ stuffWebApi/stuffWebApi/appsettings.json
@@ -1,0 +1,18 @@
+﻿{
+  "ConnectionStrings": {
+    "ChatDatabase": "Server=(localdb)\\mssqllocaldb;Database=stuffWebApi;Trusted_Connection=True;MultipleActiveResultSets=true"
+  },
+    "Logging": {
+      "IncludeScopes": false,
+      "Debug": {
+        "LogLevel": {
+          "Default": "Warning"
+        }
+      },
+      "Console": {
+        "LogLevel": {
+          "Default": "Warning"
+        }
+      }
+    }
+  }
