--- stuffioAPI/stuffioAPI/appsettings.json
@@ -1,18 +1,0 @@
-﻿{
-  "ConnectionStrings": {
-    "ChatDatabase": "Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=stuffio;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=True;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
-  },
-  "Logging": {
-    "IncludeScopes": false,
-    "Debug": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    },
-    "Console": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    }
-  }
-}
